+++ src/routes/addresses/index.ts
@@ -3,6 +3,7 @@ import * as QueryTypes from '../../types/queries/addresses';
 import * as ResponseTypes from '../../types/responses/addresses';
 import { getDbSync } from '../../utils/database';
 import { getSchemaForEndpoint } from '@blockfrost/openapi';
+import { validation } from '@blockfrost/blockfrost-utils';
 import { getAdditionalParametersFromRequest } from '../../utils/string-utils';
 import { handle400Custom, handle404, handleInvalidAddress } from '../../utils/error-handler';
 import { getAddressTypeAndPaymentCred, paymentCredToBech32Address } from '../../utils/validation';
@@ -127,17 +128,23 @@ async function addresses(fastify: FastifyInstance) {
         }
 
         const assetsAmount: ResponseTypes.AmountExtended = [];
-        // add off-chain data to all assets if they exist
 
+        // add off-chain data to all assets if they exist
         if (rows[0].amount) {
           for (const asset of rows[0].amount) {
-            const registryData = await fetchAssetMetadata(asset.unit);
+            const unit = `${asset.policy_id}${asset.asset_name}`;
+            const registryData = await fetchAssetMetadata(unit);
+            const onchainMetadata = validation.getOnchainMetadata(
+              asset.onchain_metadata,
+              asset.asset_name,
+              asset.policy_id,
+            );
 
             assetsAmount.push({
-              unit: asset.unit,
+              unit: unit,
               quantity: asset.quantity,
               decimals: registryData?.decimals ?? null,
-              has_nft_onchain_metadata: asset.has_nft_onchain_metadata,
+              has_nft_onchain_metadata: onchainMetadata !== null,
             });
           }
         }