+++ src/types/queries/assets.ts
@@ -1,4 +1,4 @@
-import { Order } from '../common';
+import { OnchainMetadata, Order } from '../common';
 export type { ResultFound } from '../common';
 
 export interface AssetsPolicyFound {
@@ -38,34 +38,6 @@ export interface Assets {
   quantity: string;
 }
 
-interface Metadata {
-  name: string;
-  description: string;
-  ticker: string | null;
-  url: string | null;
-  logo: string | null;
-}
-
-interface OnchainMetadata {
-  name: string | null;
-  image: string | null;
-}
-
-export interface Asset {
-  asset: string;
-  policy_id: string;
-  asset_name: string | null;
-  asset_name_UTF8: string | null;
-  fingerprint: string;
-  quantity: string;
-  initial_mint_tx_id: string;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  onchain_metadata: OnchainMetadata | null | any;
-  metadata: Metadata | null;
-  initial_mint_tx_hash: string;
-  mint_or_burn_count: number;
-}
-
 export interface AssetHistory {
   tx_hash: string;
   amount: string;
@@ -91,4 +63,25 @@ export interface PolicyId {
 export interface AssetAddresses {
   address: string;
   quantity: string;
+}
+
+interface Metadata {
+  name: string;
+  description: string;
+  ticker: string | null;
+  url: string | null;
+  logo: string | null;
+}
+
+export interface Asset {
+  asset: string;
+  policy_id: string;
+  asset_name: string | null;
+  fingerprint: string;
+  quantity: string;
+  initial_mint_tx_id: string;
+  onchain_metadata: OnchainMetadata | null;
+  metadata: Metadata | null;
+  initial_mint_tx_hash: string;
+  mint_or_burn_count: number;
 }